Golden Doge Solana $GDoge
We are meme coin that takes different approach from another meme coin. We don’t only create a common meme token, we create a social meme token that is intended to contribute in social and art movement. The token is called $GDoge. We use Solana blockchain as our core blockchain